starting to factor out search grid component
