+++ CRBM-Viz/src/app/Modules/simulate/browse/browse.component.html
@@ -1,27 +1,6 @@
 <div class="full-width-container">
   <h1>Simulations</h1>
   <div class="content-container">
-    <ag-grid-angular #searchGrid
-      class="ag-theme-balham search-grid"
-      [icons]="icons"
-      [frameworkComponents]="frameworkComponents"
-      [defaultColDef]="defaultColDef"
-      [columnDefs]="columnDefs"
-      [sideBar]="sideBar"
-      [statusBar]="statusBar"
-      [rowData]="rowData"
-      (gridReady)="onGridReady($event)"
-      (firstDataRendered)="onFirstDataRendered($event)"
-      (columnVisible)="onColumnVisible($event)"
-      (gridSizeChanged)="onGridSizeChanged($event)"
-      (toolPanelVisibleChanged)="onToolPanelVisibleChanged($event)"
-      [suppressHorizontalScroll]="true"
-      [suppressCellSelection]="true"
-      [suppressContextMenu]="true"
-      [suppressAutoSize]="true"
-      [headerHeight]="40"
-      [rowHeight]="35"
-    >
-    </ag-grid-angular>
+    <app-simulations-grid></app-simulations-grid>
   </div>
 </div>
