+++ CRBM-Viz/src/app/Shared/Components/grid/search-tool-panel-grid.component.sass
