+++ CRBM-Viz/src/app/Shared/shared.module.ts
@@ -2,18 +2,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
+import { FormsModule } from '@angular/forms';
 
-// Shared components
-import { FilterPipe } from './Pipes/filter.pipe';
-import { SearchBarComponent } from './Components/search-bar/search-bar.component';
-import { AccountMenuComponent } from './Components/account-menu/account-menu.component';
-import { LogoComponent } from './Components/logo/logo.component';
-import { NavigationComponent } from './Components/navigation/navigation.component';
-import { NavIconsComponent } from './Components/nav-icons/nav-icons.component';
-import { SidebarComponent } from './Components/sidebar/sidebar.component';
-import { DataTableComponent } from './Components/data-table/data-table.component';
-import { MaterialModule } from '../Modules/app-material.module';
+// Third pary modules
 import { AgGridModule } from 'ag-grid-angular';
+import { MaterialModule } from '../Modules/app-material.module';
+
 // FontAwesome for icons
 import {
   FontAwesomeModule,
@@ -35,12 +29,24 @@ import {
   faGoogle,
   faOrcid,
 } from '@fortawesome/free-brands-svg-icons';
+
+// Shared components
+import { FilterPipe } from './Pipes/filter.pipe';
+import { SearchBarComponent } from './Components/search-bar/search-bar.component';
+import { AccountMenuComponent } from './Components/account-menu/account-menu.component';
+import { LogoComponent } from './Components/logo/logo.component';
+import { NavigationComponent } from './Components/navigation/navigation.component';
+import { NavIconsComponent } from './Components/nav-icons/nav-icons.component';
+import { SidebarComponent } from './Components/sidebar/sidebar.component';
+import { DataTableComponent } from './Components/data-table/data-table.component';
 import { AlertComponent } from './Components/alert/alert.component';
 import { CallbackComponent } from './Components/callback/callback.component';
 import { HomeComponent } from './Components/home/home.component';
 import { FourComponent } from './Components/four/four.component';
 import { UnderConstructionComponent } from './Components/under-construction/under-construction.component';
-import { FormsModule } from '@angular/forms';
+import { IdRendererGridComponent } from './Components/grid/id-renderer-grid.component';
+import { SearchToolPanelGridComponent } from './Components/grid/search-tool-panel-grid.component';
+import { SimulationsGridComponent } from './Components/simulations-grid/simulations-grid.component';
 
 @NgModule({
   declarations: [
@@ -57,13 +63,16 @@ import { FormsModule } from '@angular/forms';
     HomeComponent,
     FourComponent,
     UnderConstructionComponent,
+    IdRendererGridComponent,
+    SearchToolPanelGridComponent,
+    SimulationsGridComponent,
   ],
   imports: [
     CommonModule,
     MaterialModule,
     RouterModule,
     FontAwesomeModule,
-    AgGridModule.withComponents([]),
+    AgGridModule.withComponents([IdRendererGridComponent, SearchToolPanelGridComponent]),
     FormsModule,
   ],
   exports: [
@@ -81,6 +90,7 @@ import { FormsModule } from '@angular/forms';
     HomeComponent,
     FourComponent,
     UnderConstructionComponent,
+    SimulationsGridComponent,
   ],
   entryComponents: [AlertComponent],
 })
