+++ CRBM-Viz/src/app/app.component.spec.ts
@@ -2,7 +2,6 @@ import { TestBed, async } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { BrowserDynamicTestingModule } from '@angular/platform-browser-dynamic/testing';
 import { AppComponent } from './app.component';
-import { DetailsRouteRendererComponent } from './Modules/simulate/browse/details-route-renderer.component';
 
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { AuthService } from './Shared/Services/auth0.service';
@@ -11,12 +10,12 @@ describe('AppComponent', () => {
     TestBed.configureTestingModule({
       imports: [RouterTestingModule],
       providers: [AuthService],
-      declarations: [AppComponent, DetailsRouteRendererComponent],
+      declarations: [AppComponent],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
     }).compileComponents();
     TestBed.overrideModule(BrowserDynamicTestingModule, {
       set: {
-        entryComponents: [DetailsRouteRendererComponent],
+        entryComponents: [],
       },
     });
   }));
