+++ CRBM-Viz/src/styles.scss
@@ -238,7 +238,7 @@ p:last-child {
         padding-bottom: 2px;
     }
 
-    .search-tool-panel-icon {
+    .search-tool-panel-grid-icon {
         font-size: 16px;
         width: 16px;
         height: 16px;
